De-virtualise Objects (#8)
Reading and writing the database is done via either the generated classes, or using lower-level code during serialisation/deserialisation.

This PR simplifies the class structure and type information to remove the need for virtual classses.
An `Object` instance can be used to access all types (Object, Array, ObjectArray and Store). It contains these three pointers:

1. Type information (`ObjectInfo`)
2. Parent `Object` (null for the top-level `Store` instance)
3. Data (which lives in the `Store`)

Objects are created on the stack so `unique_ptr` is no longer necessary.