Implement store sharing / locking (#17)
This PR implements store locking as discussed in #12. Also started a test application with a Sharing module.

We now generate four types of templated Objects:

- `Contained` object with `get` methods (const)
- Outer object instantiated by application code, contains Store (const) and inherits from Contained object
- Updater object with `set` methods, inherits from Contained object
- Outer Updater object instantiated by application code, contains Store and inherits from Updater object

Objects are now read-only and have separate `Updater` types to handle writing.
These call `commit` in their destructor, but applications can call it directly if they need to check for success/failure.

Updates must be consistent and not be reflected in other readers until committed.
A web client reading the database using a streaming `Reader` must always receive a consistent set of data for each store.
For efficiency, multiple readers share the same Store instance so when `update` is called on one of them then a copy of the Store is made internally. Rather than reloading from storage, a copy constructor has been added to `Store` which makes this operation more efficient.

The `Basic_Config` sample has been updated with a basic demonstration of streaming updates. This accepts a POST request with Content-Type: application/json.

Note that, as discussed, the update behaviour has been changed to 'simplest' mode:

- Stores are *not* cleared
- Incoming data overwrites values
- Arrays are overwritten, so individual values cannot be updated

TODO:

* [x] Gracefully fail 'set' and 'commit' operations if store isn't opened for write access.
* [x] Remove set/commit from generated objects and provide separate `Updater` classes.
* [x] Block or `delete` inherited methods which allow write access
* [x] Implement asynchronous updates for classes
* [x] Implement asynchronous updates for streamers. For now, report any failures back to web client so it can retry.
* [x] Tidy sample code
* [x] Review comments and documentation
* [x] Fix array value assignments when using `operator[]`