Add Union support (#25)
This PR adds support for unions and for defined types using JSONSchema `$ref`.

https://json-schema.org/understanding-json-schema/structuring#dollarref

This allows types to be defined within the `$defs` section of the schema and re-used.

This is leveraged to support `Union` types via the `oneOf` schema keyword.
The `test` application contains an example of this in the `test-config-union.cfgdb` schema. It is used in the `Updates` test module.

## Union structures

Like a regular C++ `union`, a `ConfigDB::Union` object has one or more object types overlaid in the same storage space. The size of the object is therefore governed by the size of the largest type stored. A `uint8_t` property tag indicates which type is stored.

The code generator produces a `get` method for each type of object which can be stored. The application is responsible for checking which type is present via `getTag`; if the wrong method is called, a runtime assertion will be generated.

The corresponding Union Updater class has a `setTag` method. This changes the stored object type and initialises it to default values. This is done even if the tag value doesn't change so can be used to 'reset' an object to defaults.


## Store child objects as properties

Instead of storing a separate list of `ObjectInfo` pointers for contained objects, these are now included in the object properties `PropertyInfo` list. This change is necessary because a type may be used in several places with different names.