Add `resetXXXX()` methods to updater (#46)
Objects and arrays can be reset (PR #37) but as per https://github.com/mikee47/ConfigDB/discussions/20#discussioncomment-10471624 there is currently no specific way to reset property values to their default. This PR adds `resetXXXX()` methods to generated updater code for this purpose.

Objects have a defined `defaultData` structure which contains the default values and can be useful for introspection. These could be used to reset default values, but the types do not always agree with those in the API and doing it this way is a bit clunky.