Support external schemas (#54)
This PR addresses the issues raised in #53, reworking the code generator, There are no changes to the code framework itself, and the generated classes *should* be unchanged.

The code generator parses the schema into an internal object representation which has been revised to more closely resemble that of the C++ classes. The major change is that a `Property` class can now contain an object reference, which allows the same object definition to be used with different names in various places.

The offical rules for managing schema identification and resolution is covered here https://json-schema.org/understanding-json-schema/structuring. This PR implements a simpler version which assumes that all schemas are located in the same project. The code generator parses all `.cfgdb` files together to support cross-referencing between them.

Rather than require an explicit `$id` annotation, the file name (without extension) is uses as the `schema_id`. For an example, see the `test_external.cfgdb` schema in the test application.

## Union options

Unions can now override the name of option objects usign the `title` keyword. This allows the same object type to be used more than once for the same union option.

Also, option definitions can be provided directly, in which case `title` is required.

## Validation

Checks have been added to ensure that property names are unique within their defined scope, and that generated identifiers for these are valid and also unique. For example, `value+` and `value-` produce the same identifier `value` so are not valid.

## Restrictions and limitations

- Nested references are not supported. That is, a `$ref` must point to an actual definition and not just another `$ref`.
- References can point anywhere in a schema, but following the JSON Schema convention and putting re-useable definitions in `$defs` is recommended.
- Not all possible combinations and permutations have been tested. If a particular schema does not build, or generates compile errors, then consider raising an issue so the generator can be improved.